fix for printf_chk

+++ 6_Language/1_Lexical_elements/3_Constants/4_Character_constants/1_ok/test008.c
@@ -4,6 +4,6 @@
 int main()
 {
   wchar_t wc = L'”\';
-  printf("%x\n", wc);
+  printf("%x\n", (int)wc);
   return 0;
 }
