+++ 6_Language/1_Lexical_elements/3_Constants/4_Character_constants/1_ok/test010.c
@@ -4,7 +4,7 @@
 int main()
 {
   wchar_t wc = L'怒';
-  printf("%x\n", wc);
+  printf("%x\n", (int)wc);
   return 0;
 }
